requirements.txt
streamlit
opencv-python
pytesseract
numpy
sympy
pillow
